+++ containers/index.ts
@@ -1,1 +1,1 @@
-export * from './Home';
+export * from './Landing';
